- Changed gtm_createCGPath to gtm_cgPath in GTMNSBezier+CGPath. The path   returned is now autoreleased so you don't need to worry about releasing it.
- Made some changes to the GTMNSObject+UnitTesting APIs. Specifically renamed
  gtm_createUnitTestImage to gtm_unitTestImage. The value it returns is now
  autoreleased, so no need to release it. Also change 
  gtm_createUnitTestBitmapOfSize:withData: to a C function.

- Cleaned up GTM so that it passes the Clang checker without any warnings.

+++ trunk/AppKit/GTMHotKeyTextField.m
@@ -209,7 +209,7 @@ static CFStringRef kGTM_TISPropertyUnicodeKeyLayoutData = NULL;
 
 #if MAC_OS_X_VERSION_MIN_REQUIRED >= MAC_OS_X_VERSION_10_5
 
-- (int)integerValue {
+- (NSInteger)integerValue {
   
   // Defeating NSControl
   _GTMDevAssert(NO, @"Hot key fields don't take numbers.");
